+++ src/pyiem/plot/__init__.py
@@ -1,7 +1,19 @@
 """Plotting"""
+# This was a bad idea, but we are stuck with it for now.
 
-from pyiem.plot.colormaps import *  # noqa
-from pyiem.plot.geoplot import *  # noqa
-from pyiem.plot.calendarplot import *  # noqa
-from pyiem.plot.util import centered_bins, pretty_bins  # noqa
-from pyiem.plot.layouts import figure, figure_axes  # noqa
+from .calendarplot import calendar_plot  # noqa
+from .colormaps import (
+    get_cmap,  # noqa
+    nwsprecip,  # noqa
+    nwssnow,  # noqa
+)
+from .geoplot import (
+    MapPlot,  # noqa
+)
+from .layouts import figure, figure_axes  # noqa
+from .util import (
+    centered_bins,  # noqa
+    fitbox,  # noqa
+    pretty_bins,  # noqa
+    ramp2df,  # noqa
+)