+++ tests/plot/test_calendarplot.py
@@ -2,7 +2,7 @@
 import datetime
 
 import pytest
-from pyiem.plot import calendar_plot
+from pyiem.plot.calendarplot import calendar_plot
 
 PAIN = 0.02  # how much do we care, sigh.